delete: main template
