+++ pages/archives/_ym.vue
@@ -1,14 +1,12 @@
 <template>
-  <main-template>
-    <div class="main">
-      <calendar :items="archives" :ym="$route.params.ym" />
-      <google-adsense
-        ad-slot="5919567639"
-        ad-format="auto"
-        :ad-style="{ display: 'block' }"
-      />
-    </div>
-  </main-template>
+  <div class="main">
+    <calendar :items="archives" :ym="$route.params.ym" />
+    <google-adsense
+      ad-slot="5919567639"
+      ad-format="auto"
+      :ad-style="{ display: 'block' }"
+    />
+  </div>
 </template>
 
 <script lang="ts">
@@ -16,13 +14,11 @@ import Vue from 'vue'
 
 import { fetchPost } from '~/api/blog'
 
-const MainTemplate = () => import('~/components/MainTemplate.vue')
 const GoogleAdsense = () => import('~/components/GoogleAdsense.vue')
 const Calendar = () => import('~/components/Calendar.vue')
 
 export default Vue.extend({
   components: {
-    MainTemplate,
     GoogleAdsense,
     Calendar
   },
