+++ pages/tag/_name.vue
@@ -1,23 +1,21 @@
 <template>
-  <main-template>
-    <div class="main">
-      <div class="tag-text">
-        <div class="tag">
-          {{ $route.params.name }}
-        </div>
-      </div>
-
-      <div class="top">
-        <div v-for="post in posts" :key="post.fields.title" class="items">
-          <nuxt-link
-            :to="{ name: 'posts-slug', params: { slug: post.fields.slug } }"
-          >
-            <post-card :post="post" />
-          </nuxt-link>
-        </div>
+  <div class="main">
+    <div class="tag-text">
+      <div class="tag">
+        {{ $route.params.name }}
       </div>
     </div>
-  </main-template>
+
+    <div class="top">
+      <div v-for="post in posts" :key="post.fields.title" class="items">
+        <nuxt-link
+          :to="{ name: 'posts-slug', params: { slug: post.fields.slug } }"
+        >
+          <post-card :post="post" />
+        </nuxt-link>
+      </div>
+    </div>
+  </div>
 </template>
 
 <script lang="ts">
@@ -25,12 +23,10 @@ import Vue from 'vue'
 
 import { fetchPostsByTag } from '~/api/blog'
 
-const MainTemplate = () => import('~/components/MainTemplate.vue')
 const PostCard = () => import('~/components/PostCard.vue')
 
 export default Vue.extend({
   components: {
-    MainTemplate,
     PostCard
   },
   async asyncData({ route }) {
