+++ tests/history-tags.spec.ts
@@ -1,13 +1,13 @@
 import { mount } from '@vue/test-utils'
 
-import MainTemplate from '~/components/MainTemplate.vue'
+import HistoryTags from '../components/HistoryTags.vue'
 
-describe('MainTemplate', () => {
+describe('HistoryTags', () => {
   test('correctly render', () => {
     const props = {
       tags: ['Vue']
     }
-    const wrapper = mount(MainTemplate, {
+    const wrapper = mount(HistoryTags, {
       propsData: props
     })
     expect(wrapper.exists()).toBe(true)
